requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8

celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
eventlet==0.39.0

isort==6.0.0
python-decouple==3.8
